refactor: split ATExecAlterConstrRecurse()
When changing a NOT ENFORCED foreign key constraint to ENFORCED, we
need to create the necessary triggers for enforcing the constraint. We
can set the deferrability at the time of creation and skip the code
used for other constraint alteration operations.

Additionally, move the code that iterates over each child constraint
of the constraint being altered into a separate function.