pg_verifybackup: Rename the wal-directory switch to wal-path
Future patches to pg_waldump will enable it to decode WAL directly
from tar files. This means you'll be able to specify a tar archive
path instead of a traditional WAL directory.

To keep things consistent and more versatile, we should also
generalize the input switch for pg_verifybackup. It should accept
either a directory or a tar file path that contains WALs. This change
will also aligning it with the existing manifest-path switch naming.