WIP-pg_waldump: Remove the restriction on the order of archived WAL files.
With previous patch, pg_waldump would stop decoding if WAL files were
not in the required sequence. With this patch, decoding will now
continue.  Any WAL file that is out of order will be written to a
temporary location, from which it will be read later. Once a temporary
file has been read, it will be removed.

TODO:
  Timeline switching is not handled correctly, especially when a
  timeline change occurs on the next WAL file that was previously
  written to a temporary location.